Introduce -Wall flag, and fix compiler warnings

+++ keyvi/include/keyvi/stringdistance/needleman_wunsch.h
@@ -177,7 +177,10 @@ class NeedlemanWunsch final {
     return intermediate_score;
   }
 
-  int GetScore() const { return distance_matrix_.Get(latest_calculated_row_, distance_matrix_.Columns() - 1); }
+  // TODO(Amit) Refactor DistanceMatrix and remove type cast.
+  size_t GetScore() const {
+    return static_cast<size_t>(distance_matrix_.Get(latest_calculated_row_, distance_matrix_.Columns() - 1));
+  }
 
   std::string GetCandidate() {
     std::vector<unsigned char> utf8result;
